+++ backend/src/main/java/com/ssafy/fullcourse/domain/sharefullcourse/repository/SharedFCLikeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SharedFCLikeRepository extends JpaRepository<SharedFCLike,Integer> {
+public interface SharedFCLikeRepository extends JpaRepository<SharedFCLike,Long> {
 
     SharedFCLike findByUser_UserIdAndSharedFullCourse_SharedFcId(Long userId, Long shareFcdId);
